+++ src/engine/Graphics/Context/debug-text.ts
@@ -1,5 +1,8 @@
-import { ExcaliburGraphicsContext, ImageSource, SpriteFont, SpriteSheet } from '..';
-import { Vector } from '../..';
+import { ExcaliburGraphicsContext } from '../Context/ExcaliburGraphicsContext';
+import { ImageSource } from '../ImageSource';
+import { SpriteFont } from '../SpriteFont';
+import { SpriteSheet } from '../SpriteSheet';
+import { Vector } from '../../Math/vector';
 import debugFont from './debug-font.png';
 
 /**
