+++ src/engine/TileMap/IsometricMap.ts
@@ -1,4 +1,10 @@
-import { BodyComponent, BoundingBox, Collider, ColliderComponent, CollisionType, CompositeCollider, vec, Vector } from '..';
+import { BodyComponent } from '../Collision/BodyComponent';
+import { BoundingBox} from '../Collision/BoundingBox';
+import { ColliderComponent } from '../Collision/ColliderComponent';
+import { Collider } from '../Collision/Colliders/Collider';
+import { CollisionType } from '../Collision/CollisionType'
+import { CompositeCollider } from '../Collision/Colliders/CompositeCollider';
+import { vec, Vector } from '../Math/vector';
 import { TransformComponent } from '../EntityComponentSystem/Components/TransformComponent';
 import { Entity } from '../EntityComponentSystem/Entity';
 import { DebugGraphicsComponent, ExcaliburGraphicsContext, Graphic, GraphicsComponent } from '../Graphics';
